Provide more information in a QiskitResult
Previously, QiskitResult would compute qubit count, depth, and job IDs
only on demand.  Now, it returns qubit count and depth directly and
provides a means of acquiring job ID.  Specifically, it tags all jobs
with a unique tag, which it returns via the job_tags attribute.  The
provider then can be queried for jobs using the unique tag as a
filter.